requirements.txt
annotated-types==0.7.0
bcrypt==5.0.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
cryptography==45.0.5
dnspython==2.8.0
docx2txt==0.9
email-validator==2.3.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fsspec==2025.7.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-speech==2.33.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.3.2
grpcio==1.72.1
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
openai==1.84.0
openai-whisper==20250625
pdfkit==1.0.0
pdfminer.six==20250506
pdfplumber==0.11.7
phonenumbers==9.0.5
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyjokes==0.8.3
pyparsing==3.2.3
pypdfium2==4.30.1
pypiwin32==223
python-dotenv==1.1.1
pyttsx3==2.98
pywin32==310
regex==2024.11.6
reportlab==4.4.2
requests==2.32.3
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.48
sympy==1.14.0
textblob==0.19.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
Werkzeug==3.1.3
wikipedia==1.4.0
WTForms==3.2.1

gunicorn==22.0.0
psycopg2-binary==2.9.9
PyPDF2==3.0.1
scikit-learn==1.3.2
